_DJANGO_APP/requirements.txt
alabaster==0.7.10
dj-database-url==0.5.0
Django==2.0.5
django-bootstrap3==11.0.0
django-datatable==0.3.1
django-filter==2.0.0
django-import-export==1.0.1
django-model-utils==3.1.2
django-tables2==1.21.2
djangorestframework==3.8.2
djangorestframework-datatables==0.4.0
gunicorn==19.9.0
psycopg2==2.7.5
tables==3.4.3
tablib==0.12.1
Werkzeug==0.14.1
whitenoise==4.0
